+++ packages/example/next.config.js
@@ -1,9 +1,15 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
+  reactStrictMode: true,
+  webpack: (config) => {
+    config.externals.push('pino-pretty', 'lokijs', 'encoding');
+    return config;
+  },
   i18n: {
     defaultLocale: 'en-US',
     locales: [
       'ar-AR',
+      'de-DE',
       'en-US',
       'es-419',
       'fr-FR',
@@ -16,10 +22,10 @@ module.exports = {
       'th-TH',
       'tr-TR',
       'uk-UA',
+      'vi-VN',
       'zh-CN',
       'zh-HK',
       'zh-TW',
     ],
   },
-  reactStrictMode: true,
 };