+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -37,6 +37,7 @@ import { okxWallet } from './okxWallet/okxWallet';
 import { omniWallet } from './omniWallet/omniWallet';
 import { oneInchWallet } from './oneInchWallet/oneInchWallet';
 import { oneKeyWallet } from './oneKeyWallet/oneKeyWallet';
+import { paraSwapWallet } from './paraSwapWallet/paraswapWallet';
 import { phantomWallet } from './phantomWallet/phantomWallet';
 import { rabbyWallet } from './rabbyWallet/rabbyWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
@@ -55,6 +56,7 @@ import { trustWallet } from './trustWallet/trustWallet';
 import { uniswapWallet } from './uniswapWallet/uniswapWallet';
 import { valoraWallet } from './valoraWallet/valoraWallet';
 import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
+import { wigwamWallet } from './wigwamWallet/wigwamWallet';
 import { xdefiWallet } from './xdefiWallet/xdefiWallet';
 import { zealWallet } from './zealWallet/zealWallet';
 import { zerionWallet } from './zerionWallet/zerionWallet';
@@ -99,6 +101,7 @@ export {
   omniWallet,
   oneInchWallet,
   oneKeyWallet,
+  paraSwapWallet,
   phantomWallet,
   rabbyWallet,
   rainbowWallet,
@@ -117,6 +120,7 @@ export {
   uniswapWallet,
   valoraWallet,
   walletConnectWallet,
+  wigwamWallet,
   xdefiWallet,
   zealWallet,
   zerionWallet,