+++ site/next.config.js
@@ -8,10 +8,15 @@ const withVanillaExtract = createVanillaExtractPlugin();
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  webpack: (config) => {
+    config.externals.push('pino-pretty', 'lokijs', 'encoding');
+    return config;
+  },
   i18n: {
     locales: [
       'en-US',
       'ar',
+      'de',
       'es-419',
       'fr',
       'hi',
@@ -23,6 +28,7 @@ const nextConfig = {
       'th',
       'tr',
       'ua',
+      'vi',
       'zh-CN',
       'zh-HK',
       'zh-TW',